+++ apiron/service/discoverable.py
@@ -1,12 +1,12 @@
-from apiron.service.base import Service
+from apiron.service.base import ServiceBase
 
 
-class DiscoverableService(Service):
+class DiscoverableService(ServiceBase):
     """
     A Service whose hosts are determined via a host resolver.
     A host resolver is any class with a :func:`resolve` method
-    that takes a service name as its sole argument and returns a
-    list of host names that correspond to that service.
+    that takes a service name as its sole argument
+    and returns a list of host names that correspond to that service.
     """
 
     @classmethod
@@ -14,7 +14,12 @@ class DiscoverableService(Service):
         return cls.host_resolver_class.resolve(cls.service_name)
 
     def __str__(self):
-        return str(self.__class__)
+        return self.service_name
 
     def __repr__(self):
-        return repr(self.__class__)
+        klass = self.__class__
+        return '{klass}(service_name={service_name}, host_resolver={host_resolver})'.format(
+            klass=klass.__name__,
+            service_name=klass.service_name,
+            host_resolver=klass.host_resolver_class.__name__,
+        )