+++ tests/service/test_base.py
@@ -1,6 +1,6 @@
 import pytest
 
-from apiron import Service
+from apiron import Service, ServiceBase
 
 
 @pytest.fixture(scope='class')
@@ -8,6 +8,13 @@ def service():
     class SomeService(Service):
         domain = 'http://foo.com'
     return SomeService
+
+class TestServiceBase:
+    def test_get_hosts_returns_empty_list_by_default(self):
+        assert [] == ServiceBase.get_hosts()
+
+    def test_required_headers_returns_empty_dict_by_default(self, service):
+        assert {} == service.required_headers
 
 
 class TestService:
@@ -26,5 +33,5 @@ class TestService:
     def test_repr_method_on_instance(self, service):
         assert 'SomeService(domain=http://foo.com)' == repr(service())
 
-    def test_required_hosts_returns_dictionary(self, service):
+    def test_required_headers_returns_empty_dict_by_default(self, service):
         assert {} == service.required_headers